requirements.txt
discord
discord.py
logging
python-dotenv